+++ bht_web.py
@@ -1,7 +1,3 @@
-import os
 from web import create_app
 
-# FLASK_ENV is deprecated
-# Set BHT_ENV  to choose your configuration
-bht_env = os.environ.get("BHT_ENV")
-app = create_app(bht_env)
+app = create_app()