+++ web/app_logger.py
@@ -42,7 +42,6 @@ class AppLogger:
                 _app.logger.addHandler(mail_handler)
 
         if _app.config.get("LOG_TO_STDOUT") and _app.config["LOG_TO_STDOUT"]:
-            print("STDO")
             stream_handler = logging.StreamHandler()
             stream_handler.setLevel(logging.INFO)
             stream_handler.setFormatter(
@@ -52,7 +51,6 @@ class AppLogger:
             )
             _app.logger.addHandler(stream_handler)
         else:
-            print("FILE")
             # log INFO to file
             if not os.path.exists("logs"):
                 os.mkdir("logs")